// src/app/modules/program/models/program.ts
export interface Program {
 programId: string;
 programUmrahYear: number;
 programTransDetails: string;
 programPrice: number;
 programPathID: string;
 programName: string;
 programDesc: string;
 programCategoryID: string;
 progAirTicketIncluded: boolean;
 lunchPrice: number;
 lunchIncluded: boolean;
 lunchDetails: string;
 lunchAvailable: boolean;
 internalTransportations: string;
 hotelStars: string;
 hotelNights: string;
 hotelName: string;
 distanceFromHaram: number;
 dinnerPrice: number;
 dinnerIncluded: boolean;
 dinnerDetails: string;
 dinnerAvailable: boolean;
 breakfastPrice: number;
 breakfastIncluded: boolean;
 breakfastDetails: string;
 breakfastAvailable: string;
 airlineID: string;
 airlineDesc: string;
 ElGenValley: boolean;
 HaramenSharefen: boolean;
 Hira: boolean;
 MBreakfastAvailable: boolean;
 MBreakfastDetails: string;
 MBreakfastIncluded: boolean;
 MBreakfastPrice: string;
 MDinnerAvailable: boolean;
 MDinnerDetails: string;
 MDinnerIncluded: boolean;
 MDinnerPrice: number;
 MDistanceFromHaram: number;
 MHotelName: string;
 MHotelNights: string;
 MHotelStars: string;
 MLunchAvailable: boolean;
 MLunchDetails: string;
 MLunchIncluded: boolean;
 MLunchPrice: number;
 MaccaMusieum: boolean;
 MasjedZuElqeblaten: boolean;
 MoulaTomb: boolean;
 NourMountain: boolean;
 PrinitHolyQuran: boolean;
 Qebaa: boolean;
 SevenMasjed: boolean;
 Thour: boolean;
 Uhod: boolean;
}